+++ ntex/src/client/h2proto.rs
@@ -19,11 +19,12 @@ pub(super) async fn send_request(
     body: Body,
     timeout: Millis,
 ) -> Result<(ResponseHead, Payload), ClientError> {
-    // log::trace!(
-    //     "{}: Sending client request: {req:?} {:?}",
-    //     client.client.tag(),
-    //     body.size()
-    // );
+    #[cfg(feature = "trace")]
+    log::trace!(
+        "{}: Sending client request: {req:?} {:?}",
+        client.client.tag(),
+        body.size()
+    );
     let length = body.size();
     let eof = if req.head.method == Method::HEAD {
         true
@@ -113,11 +114,12 @@ async fn get_response(
             headers,
             eof,
         } => {
-            // log::trace!(
-            //     "{}: {:?} got response (eof: {eof}): {pseudo:#?}\nheaders: {headers:#?}",
-            //     stream.tag(),
-            //     stream.id(),
-            // );
+            #[cfg(feature = "trace")]
+            log::trace!(
+                "{}: {:?} got response (eof: {eof}): {pseudo:#?}\nheaders: {headers:#?}",
+                stream.tag(),
+                stream.id(),
+            );
 
             match pseudo.status {
                 Some(status) => {
@@ -128,11 +130,12 @@ async fn get_response(
                     let payload = if eof {
                         Payload::None
                     } else {
-                        // log::debug!(
-                        //     "{}: Creating local payload stream for {:?}",
-                        //     stream.tag(),
-                        //     stream.id()
-                        // );
+                        #[cfg(feature = "trace")]
+                        log::debug!(
+                            "{}: Creating local payload stream for {:?}",
+                            stream.tag(),
+                            stream.id()
+                        );
                         let (pl, payload) = H2Payload::create(stream.empty_capacity());
 
                         crate::rt::spawn(async move {
@@ -154,20 +157,22 @@ async fn get_response(
 
                                 match kind {
                                     h2::MessageKind::Data(data, cap) => {
-                                        // log::trace!(
-                                        //     "{}: Got data chunk for {:?}: {:?}",
-                                        //     stream.tag(),
-                                        //     stream.id(),
-                                        //     data.len()
-                                        // );
+                                        #[cfg(feature = "trace")]
+                                        log::trace!(
+                                            "{}: Got data chunk for {:?}: {:?}",
+                                            stream.tag(),
+                                            stream.id(),
+                                            data.len()
+                                        );
                                         pl.feed_data(data, cap);
                                     }
                                     h2::MessageKind::Eof(item) => {
-                                        // log::trace!(
-                                        //     "{}: Got payload eof for {:?}: {item:?}",
-                                        //     stream.tag(),
-                                        //     stream.id(),
-                                        // );
+                                        #[cfg(feature = "trace")]
+                                        log::trace!(
+                                            "{}: Got payload eof for {:?}: {item:?}",
+                                            stream.tag(),
+                                            stream.id(),
+                                        );
                                         match item {
                                             h2::StreamEof::Data(data) => {
                                                 pl.feed_eof(data);
@@ -181,10 +186,11 @@ async fn get_response(
                                         }
                                     }
                                     h2::MessageKind::Disconnect(err) => {
-                                        // log::trace!(
-                                        //     "{}: Connection is disconnected {err:?}",
-                                        //     stream.tag(),
-                                        // );
+                                        #[cfg(feature = "trace")]
+                                        log::trace!(
+                                            "{}: Connection is disconnected {err:?}",
+                                            stream.tag(),
+                                        );
                                         pl.set_error(
                                             io::Error::new(
                                                 io::ErrorKind::UnexpectedEof,
@@ -229,12 +235,13 @@ async fn send_body(
     loop {
         match poll_fn(|cx| body.poll_next_chunk(cx)).await {
             Some(Ok(b)) => {
-                // log::trace!(
-                //     "{}: {:?} sending chunk, {} bytes",
-                //     stream.tag(),
-                //     stream.id(),
-                //     b.len()
-                // );
+                #[cfg(feature = "trace")]
+                log::trace!(
+                    "{}: {:?} sending chunk, {} bytes",
+                    stream.tag(),
+                    stream.id(),
+                    b.len()
+                );
                 stream
                     .send_payload(b, false)
                     .await
@@ -242,7 +249,8 @@ async fn send_body(
             }
             Some(Err(e)) => return Err(e.into()),
             None => {
-                // log::trace!("{}: {:?} eof of send stream ", stream.tag(), stream.id());
+                #[cfg(feature = "trace")]
+                log::trace!("{}: {:?} eof of send stream ", stream.tag(), stream.id());
                 stream
                     .send_payload(Bytes::new(), true)
                     .await