+++ ntex-bytes/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## [1.6.0] (2026-04-26)
+## [1.6.0] (2026-05-02)
 
 * Add new type BytesPages and BytePage